+++ Android/app/src/main/java/com/sundaegukbap/banchango/presentation/reciperecommend/RecipeRecommendViewModel.kt
@@ -1,6 +1,7 @@
 package com.sundaegukbap.banchango.presentation.reciperecommend
 
 import androidx.lifecycle.ViewModel
+import com.sundaegukbap.banchango.model.Recipe
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
