--- Android/app/src/main/java/com/sundaegukbap/banchango/presentation/reciperecommend/RecipeRepository.kt
@@ -1,5 +1,0 @@
-package com.sundaegukbap.banchango.presentation.reciperecommend
-
-interface RecipeRepository {
-    fun getRecipeRecommendation(): List<Recipe>
-}
