+++ Android/app/src/main/java/com/sundaegukbap/banchango/repository/RecipeRepository.kt
@@ -1,0 +1,7 @@
+package com.sundaegukbap.banchango.repository
+
+import com.sundaegukbap.banchango.model.Recipe
+
+interface RecipeRepository {
+    fun getRecipeRecommendation(): List<Recipe>
+}
